% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_bib_file.R
\name{auto_bib_file}
\alias{auto_bib_file}
\title{Compiles .bib files}
\usage{
auto_bib_file(PATH1, PATH2, PATH3, PATH4)
}
\arguments{
\item{PATH1}{path to .bib directory}

\item{PATH2}{path to .bib directory}

\item{PATH3}{path to .bib directory}

\item{PATH4}{path to .bib directory}
}
\value{
a thesis.bib file stored in the bib/ directory (should be in the aggiedown project)
}
\description{
Takes filepaths where .bib files are stored and compiles them
}
