% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_coef_plot.R
\name{manual_coef_plot}
\alias{manual_coef_plot}
\title{Create a coefficient plot manually}
\usage{
manual_coef_plot(df, model_names, color_vals, shape_vals, legend_nrow)
}
\arguments{
\item{df}{a data frame of model output(s) with the following columns: Estimate, term, model}

\item{model_names}{a vector of names matching the names in the 'model' column of the df}

\item{color_vals}{a vector of color values that is the same length as the model names}

\item{shape_vals}{a vector of shape values that is the same length as the model names}

\item{legend_nrow}{a numeric value for the number of rows the legend should span}
}
\value{
A coefficient plot
}
\description{
Takes a data frame with one or more models row-binded together and creates
a coefficient plot from it. Based on jtools but takes data frame input to
be flexible to different model inputs
}
\examples{
patents <- read_patents(file_dir = "patent_files/")

}
