% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_diss_chapters.R
\name{prep_diss_chapters}
\alias{prep_diss_chapters}
\title{Convert Google Doc to Rmd}
\usage{
prep_diss_chapters(GDOC_URL, GDOC_PATH, intro = F)
}
\arguments{
\item{GDOC_URL}{the URL to the Google Doc}

\item{GDOC_PATH}{the path to a directory where the conversions will be stored}

\item{intro}{a Boolean argument to indicate whether this document is the Introduction section (e.g. has no chunks or headers embedded). Default intro = F.}
}
\value{
Four files: 01_drive_manuscript.txt, 01_drive_manuscript.docx, 02_manuscript.Rmd, 04_refs.bib
}
\description{
Takes a Google Doc with embedded elements (chunks) and converts it into an Rmd that can be used with different LaTeX templates. This was designed specifically for aggiedown, and converting Google Docs that have Zotero citations embedded via the plug-in, formatted to their BibTex format.
}
